requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
datadog-api-client==2.43.0
faiss-cpu==1.12.0
fastapi==0.117.1
h11==0.16.0
idna==3.10
numpy==2.3.3
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyGithub==2.8.1
PyJWT==2.10.1
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
slack_sdk==3.36.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
python-dotenv==1.0.1
kubernetes==33.1.0
asyncpg==0.30.0
